refactor: improve code organization and maintainability
Major improvements:
- Extract constants to dedicated file (src/constants.ts)
- Refactor server.ts (337 → 77 lines):
  * Extract schemas to src/server/schemas.ts
  * Extract response utilities to src/server/response-utils.ts
  * Extract request handlers to src/server/handlers.ts
- Refactor client-manager.ts (432 → ~380 lines):
  * Extract retry logic to src/utils/retry.ts
- Improve type safety with proper CallToolResult types
- Better code organization with single responsibility principle

Benefits:
- Improved maintainability and readability
- Easier to test individual components
- Better separation of concerns
- Reduced file sizes for easier navigation
